fix clangd warnings

+++ f4/server/services/sdo_service.cpp
@@ -65,7 +65,7 @@ void SdoService::handle_recv_frames() {
     ExpeditedSdo tsdo;
     SdoAbortCode abort_code = SdoAbortCode::general_error;
     ODEntry* dictionary_end = _server._dictionary + _server._dictionary_size;
-    ODObjectKey key = {rsdo.index, rsdo.subindex};
+    ODObjectKey key = {static_cast<uint16_t>(rsdo.index), static_cast<uint16_t>(rsdo.subindex)};
 
     const ODEntry* od_entry = emb::binary_find(_server._dictionary, dictionary_end, key);
